Purge most of the help pages
We are keeping a white listed and much smaller list of documents
that we know will be helpful and we have time to review and fixup

Some of what we are loosing here will doubtlessly be helpful.
When we have time to sift more carefully through this list we
should start bringing a lot of this stuff back.
[endlessm/eos-shell#3909]