Add some links to the video pages from the home page
We will probably want to revert this commit and do this with a new
link style in the future.

Right now this involves copying images and article titles cross
repo from getting-started-docs to here, and translating them over
again. Which is nasty and what would be solved by a new link style.

But we want the thumbs in this release, so we'll do it like this
for now.
[endlessm/eos-shell#4073]