Use Spanish content for Portuguese
Our Portuguese translation are not updated and would be half in
English. Spanish is a better substitute for now.

Symlinked the portuguses po file to point to the spanish location and
updated our tx config to update a dummy file with the portuguese translations.

This should all be reverted later.
[endlessm/eos-shell#4064]