New icons, screenshots and illustrations from design
Named things with an eos prefix to keep our assets obvious in the
big list. Also added a secondary prefix of the type of image--
control center screenshot, desktop screenshot, keyboard illustration.
So that if we say update the control center significantly it will be
obvious which assets need updates.
[endlessm/eos-shell#4235]