New English assets
Previously, Spanish assets were used in their place,
so the existing assets are moved to the es directory,
and new assets are added to the C directory.

We are missing a design-provided assets for eos-discovery-center.png,
so this asset uses a temporary image generated by the developer.

Includes new app store icon in the "add apps" page assets.

[endlessm/eos-shell#4678]
[endlessm/eos-shell#4759]