Fix translation of support email address
The links were not properly escaped for translation,
so this string was not getting translated.
For now, let's just show the email address in plain text.
If we had a better offline email client, we might want
to reconsider and fix the links to translate properly.

Changes have been propagated through Transifex.

https://phabricator.endlessm.com/T10841