Integrate zh_CN translations with Transifex
Include links to use zh_CN as fallback for missing zh translations.

Use the upstream zh_CN translations as a starting point,
but note that there are many paragraphs that show up in Transifex
as needing translation only because of a simple change
(e.g., "top" to "bottom").  It would be good to look at the diff
on this commit to establish a better baseline for these strings
with simply changes.